Prism/DataConnectors/mongodb/requirements.txt
# Asynchronous MongoDB dependencies
motor>=3.3.0
python-dotenv>=1.0.0
dnspython>=2.0.0
